IN-172 feat: add loading for all components
