+++ package.json
@@ -34,6 +34,8 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
+    "mobx": "^6.13.3",
+    "mobx-react-lite": "^4.0.7",
     "next": "14.2.7",
     "nextjs-toploader": "^3.6.15",
     "react": "^18",
