+++ src/common/components/layout/Layout.tsx
@@ -1,8 +1,10 @@
 import React, { ElementRef, ReactElement, forwardRef } from 'react'
 import { ToastContainer } from 'react-toastify'
 
-import { Header, ScrollArea } from '@/common'
+import { Header, Loader, ScrollArea } from '@/common'
+import { generalStore } from '@/common/modal/store'
 import { cn } from '@/common/utils/cn'
+import { observer } from 'mobx-react-lite'
 import NextTopLoader from 'nextjs-toploader'
 
 type Props = {
@@ -12,6 +14,8 @@ type Props = {
 
 export const Layout = forwardRef<ElementRef<'div'>, Props>(
   ({ children, className, ...rest }, ref) => {
+    const isLoading = generalStore.isLoading
+
     return (
       <div
         className={
@@ -20,6 +24,7 @@ export const Layout = forwardRef<ElementRef<'div'>, Props>(
         ref={ref}
         {...rest}
       >
+        {isLoading && <Loader />}
         <NextTopLoader
           color={'#397DF6'}
           showSpinner={false}
@@ -42,7 +47,8 @@ export const Layout = forwardRef<ElementRef<'div'>, Props>(
 )
 
 Layout.displayName = 'Layout'
+const LayoutWithStore = observer(Layout)
 
 export const getBaseLayout = (page: ReactElement) => {
-  return <Layout>{page}</Layout>
+  return <LayoutWithStore>{page}</LayoutWithStore>
 }
