+++ src/common/modal/store.ts
@@ -1,0 +1,19 @@
+import { makeAutoObservable } from 'mobx'
+
+class GeneralStore {
+  isLoading = false
+
+  constructor() {
+    makeAutoObservable(this)
+  }
+
+  turnOffLoading() {
+    this.isLoading = false
+  }
+
+  turnOnLoading() {
+    this.isLoading = true
+  }
+}
+
+export const generalStore = new GeneralStore()
