+++ src/features/auth/ui/SignUpForm.tsx
@@ -1,7 +1,6 @@
 import {
   Button,
   Card,
-  Loader,
   Typography,
   typographyVariants,
   useTranslation,
@@ -15,9 +14,10 @@ import {
   ExternalServicesRegistration,
 } from '@/features/auth'
 import { useSignUp } from '@/features/auth/model'
+import { observer } from 'mobx-react-lite'
 import Link from 'next/link'
 
-export const SignUpForm = () => {
+export const SignUpForm = observer(() => {
   const { t } = useTranslation()
   const {
     control,
@@ -32,7 +32,6 @@ export const SignUpForm = () => {
 
   return (
     <>
-      {isLoading && <Loader />}
       <Card className={'pt-6 pb-7 contents md:block'}>
         <Typography
           className={'text-center mb-3'}
@@ -119,4 +118,4 @@ export const SignUpForm = () => {
       </Card>
     </>
   )
-}
+})
